reorg-docs: split 'Features' into 'What is InvenioRDM?' and 'Use' [+]
- "What is InvenioRDM?" is targeted at prospective adopters. It
highlights at a high level with a "marketing" lens the main features or
properties of InvenioRDM.
- "Use" targets end-users and content managers/staff/site
admninistrators (from discussion, "content manager" was suggested to
refer to those with access to the administration panel and not be
ambiguous with 'admin' which has been or even is used to denote
superusers; however, with the administration panel called... the
"administration panel" we kind of expect administrators to access it 🤷).
It covers the usage of InvenioRDM.
- Placed "Use" section after the "What is InvenioRDM" (Features) section
because it shortens the distance between the "Install" and "Operate >
deploy" sections and it was at the end of the "Features" section before
so still in keeping.
- Some content that should be improved in future tickets:
  * change the image for "Setting a record to be restricted" to have the
"Restricted" choice chosen
  * more content on the use section
- Adds glightbox plugin and closes #680